+++ packages/core/hardhat.config.js
@@ -3,8 +3,7 @@ const path = require('path');
 
 require('dotenv/config');
 
-require('@nomiclabs/hardhat-ethers');
-require('@nomiclabs/hardhat-etherscan');
+require('@nomicfoundation/hardhat-ethers');
 
 for (const f of fs.readdirSync(path.join(__dirname, 'hardhat'))) {
   require(path.join(__dirname, 'hardhat', f));