+++ packages/plugin-hardhat/test/beacon-linked-libraries.js
@@ -11,7 +11,7 @@ test('without flag', async t => {
   const safeMathLib = await deployLibrary('SafeMath');
 
   // Attempting to Deploy Token
-  const Token = await getLinkedContractFactory('Token', { SafeMath: safeMathLib.address });
+  const Token = await getLinkedContractFactory('Token', { SafeMath: await safeMathLib.getAddress() });
   await t.throwsAsync(() => upgrades.deployBeacon(Token), undefined, 'Contract `Token` is not upgrade safe');
 });
 
@@ -25,7 +25,7 @@ test('with flag', async t => {
   const safePctLib = await deployLibrary('SafePercent');
 
   // Deploying Token
-  const Token = await getLinkedContractFactory('Token', { SafeMath: safeMathLib.address });
+  const Token = await getLinkedContractFactory('Token', { SafeMath: await safeMathLib.getAddress() });
   const beacon = await upgrades.deployBeacon(Token, {
     unsafeAllow: ['external-library-linking'],
   });
@@ -35,7 +35,7 @@ test('with flag', async t => {
   t.is('V1', await token.getLibraryVersion());
 
   // Deploying Token with different Library
-  const TokenNew = await getLinkedContractFactory('Token', { SafeMath: safeMathLib2.address });
+  const TokenNew = await getLinkedContractFactory('Token', { SafeMath: await safeMathLib2.getAddress() });
   const beaconNew = await upgrades.deployBeacon(TokenNew, {
     unsafeAllow: ['external-library-linking'],
   });
@@ -45,38 +45,38 @@ test('with flag', async t => {
   t.is('V2', await tokenNew.getLibraryVersion());
 
   // Attempting to upgrade to TokenV2 using same library
-  const TokenV2 = await getLinkedContractFactory('TokenV2', { SafeMath: safeMathLib.address });
+  const TokenV2 = await getLinkedContractFactory('TokenV2', { SafeMath: await safeMathLib.getAddress() });
   await upgrades.upgradeBeacon(beacon, TokenV2, {
     unsafeAllow: ['external-library-linking'],
   });
-  const token2 = TokenV2.attach(token.address);
+  const token2 = TokenV2.attach(await token.getAddress());
 
-  t.is(token.address, token2.address);
+  t.is(await token.getAddress(), await token2.getAddress());
   t.is('10000', (await token2.totalSupply()).toString());
   t.is('V1', await token2.getLibraryVersion());
 
   // Attempting to upgrade to same TokenV2 using different library
-  const TokenV2New = await getLinkedContractFactory('TokenV2', { SafeMath: safeMathLib2.address });
+  const TokenV2New = await getLinkedContractFactory('TokenV2', { SafeMath: await safeMathLib2.getAddress() });
   await upgrades.upgradeBeacon(beacon, TokenV2New, {
     unsafeAllow: ['external-library-linking'],
   });
-  const token2New = TokenV2New.attach(token.address);
+  const token2New = TokenV2New.attach(await token.getAddress());
 
-  t.is(token.address, token2New.address);
+  t.is(await token.getAddress(), await token2New.getAddress());
   t.is('10000', (await token2New.totalSupply()).toString());
   t.is('V2', await token2New.getLibraryVersion());
 
   // Attempting to upgrade to LibraryTesterV3 using multiple libraries
   const TokenV3 = await getLinkedContractFactory('TokenV3', {
-    SafeMath: safeMathLib.address,
-    SafePercent: safePctLib.address,
+    SafeMath: await safeMathLib.getAddress(),
+    SafePercent: await safePctLib.getAddress(),
   });
   await upgrades.upgradeBeacon(beacon, TokenV3, {
     unsafeAllow: ['external-library-linking'],
   });
-  const token3 = TokenV3.attach(token.address);
+  const token3 = TokenV3.attach(await token.getAddress());
 
-  t.is(token.address, token3.address);
+  t.is(await token.getAddress(), await token3.getAddress());
   t.is('10000', (await token3.totalSupply()).toString());
   t.is('V1', await token3.getLibraryVersion());
 
@@ -112,7 +112,7 @@ function linkBytecode(artifact, libraries) {
 async function deployLibrary(libraryName) {
   const Library = await ethers.getContractFactory(libraryName);
   const library = await Library.deploy();
-  await library.deployed();
+  await library.waitForDeployment();
   return library;
 }