+++ packages/plugin-hardhat/test/import.js
@@ -45,10 +45,10 @@ test('implementation happy path', async t => {
   const { GreeterProxiable } = t.context;
 
   const impl = await GreeterProxiable.deploy();
-  await impl.deployed();
+  await impl.waitForDeployment();
 
-  const contract = await upgrades.forceImport(impl.address, GreeterProxiable);
-  t.is(impl.address, contract.address);
+  const contract = await upgrades.forceImport(await impl.getAddress(), GreeterProxiable);
+  t.is(await impl.getAddress(), await contract.getAddress());
   t.is('', await contract.greet());
 
   const greeter = await upgrades.deployProxy(GreeterProxiable, ['Hello, Hardhat!'], {
@@ -70,21 +70,21 @@ test('transparent happy path', async t => {
   const { Greeter, GreeterV2, ProxyAdmin, TransparentUpgradableProxy } = t.context;
 
   const impl = await Greeter.deploy();
-  await impl.deployed();
+  await impl.waitForDeployment();
   const admin = await ProxyAdmin.deploy();
-  await admin.deployed();
+  await admin.waitForDeployment();
   const proxy = await TransparentUpgradableProxy.deploy(
-    impl.address,
-    admin.address,
+    await impl.getAddress(),
+    await admin.getAddress(),
     getInitializerData(Greeter.interface, ['Hello, Hardhat!']),
   );
-  await proxy.deployed();
+  await proxy.waitForDeployment();
 
-  const greeter = await upgrades.forceImport(proxy.address, Greeter);
+  const greeter = await upgrades.forceImport(await proxy.getAddress(), Greeter);
   t.is(await greeter.greet(), 'Hello, Hardhat!');
 
   const greeter2 = await upgrades.upgradeProxy(greeter, GreeterV2);
-  await greeter2.deployed();
+  await greeter2.waitForDeployment();
   t.is(await greeter2.greet(), 'Hello, Hardhat!');
   await greeter2.resetGreeting();
   t.is(await greeter2.greet(), 'Hello World');
@@ -94,18 +94,18 @@ test('uups happy path', async t => {
   const { GreeterProxiable, GreeterV2Proxiable, ERC1967Proxy } = t.context;
 
   const impl = await GreeterProxiable.deploy();
-  await impl.deployed();
+  await impl.waitForDeployment();
   const proxy = await ERC1967Proxy.deploy(
-    impl.address,
+    await impl.getAddress(),
     getInitializerData(GreeterProxiable.interface, ['Hello, Hardhat!']),
   );
-  await proxy.deployed();
+  await proxy.waitForDeployment();
 
-  const greeter = await upgrades.forceImport(proxy.address, GreeterProxiable);
+  const greeter = await upgrades.forceImport(await proxy.getAddress(), GreeterProxiable);
   t.is(await greeter.greet(), 'Hello, Hardhat!');
 
   const greeter2 = await upgrades.upgradeProxy(greeter, GreeterV2Proxiable);
-  await greeter2.deployed();
+  await greeter2.waitForDeployment();
   t.is(await greeter2.greet(), 'Hello, Hardhat!');
   await greeter2.resetGreeting();
   t.is(await greeter2.greet(), 'Hello World');
@@ -115,18 +115,21 @@ test('beacon proxy happy path', async t => {
   const { Greeter, GreeterV2, UpgradableBeacon, BeaconProxy } = t.context;
 
   const impl = await Greeter.deploy();
-  await impl.deployed();
-  const beacon = await UpgradableBeacon.deploy(impl.address);
-  await beacon.deployed();
-  const proxy = await BeaconProxy.deploy(beacon.address, getInitializerData(Greeter.interface, ['Hello, Hardhat!']));
-  await proxy.deployed();
+  await impl.waitForDeployment();
+  const beacon = await UpgradableBeacon.deploy(await impl.getAddress());
+  await beacon.waitForDeployment();
+  const proxy = await BeaconProxy.deploy(
+    await beacon.getAddress(),
+    getInitializerData(Greeter.interface, ['Hello, Hardhat!']),
+  );
+  await proxy.waitForDeployment();
 
-  const greeter = await upgrades.forceImport(proxy.address, Greeter);
+  const greeter = await upgrades.forceImport(await proxy.getAddress(), Greeter);
   t.is(await greeter.greet(), 'Hello, Hardhat!');
 
   await upgrades.upgradeBeacon(beacon, GreeterV2);
-  const greeter2 = GreeterV2.attach(greeter.address);
-  await greeter2.deployed();
+  const greeter2 = GreeterV2.attach(await greeter.getAddress());
+  await greeter2.waitForDeployment();
   t.is(await greeter2.greet(), 'Hello, Hardhat!');
   await greeter2.resetGreeting();
   t.is(await greeter2.greet(), 'Hello World');
@@ -136,12 +139,12 @@ test('beacon happy path', async t => {
   const { Greeter, GreeterV2, UpgradableBeacon } = t.context;
 
   const impl = await Greeter.deploy();
-  await impl.deployed();
-  const beacon = await UpgradableBeacon.deploy(impl.address);
-  await beacon.deployed();
+  await impl.waitForDeployment();
+  const beacon = await UpgradableBeacon.deploy(await impl.getAddress());
+  await beacon.waitForDeployment();
 
-  const beaconImported = await upgrades.forceImport(beacon.address, Greeter);
-  t.is(await beaconImported.implementation(), impl.address);
+  const beaconImported = await upgrades.forceImport(await beacon.getAddress(), Greeter);
+  t.is(await beaconImported.implementation(), await impl.getAddress());
 
   await upgrades.upgradeBeacon(beacon, GreeterV2);
 });
@@ -150,18 +153,18 @@ test('import proxy using contract instance', async t => {
   const { GreeterProxiable, GreeterV2Proxiable, ERC1967Proxy } = t.context;
 
   const impl = await GreeterProxiable.deploy();
-  await impl.deployed();
+  await impl.waitForDeployment();
   const proxy = await ERC1967Proxy.deploy(
-    impl.address,
+    await impl.getAddress(),
     getInitializerData(GreeterProxiable.interface, ['Hello, Hardhat!']),
   );
-  await proxy.deployed();
+  await proxy.waitForDeployment();
 
   const greeter = await upgrades.forceImport(proxy, GreeterProxiable);
   t.is(await greeter.greet(), 'Hello, Hardhat!');
 
   const greeter2 = await upgrades.upgradeProxy(greeter, GreeterV2Proxiable);
-  await greeter2.deployed();
+  await greeter2.waitForDeployment();
   t.is(await greeter2.greet(), 'Hello, Hardhat!');
   await greeter2.resetGreeting();
   t.is(await greeter2.greet(), 'Hello World');
@@ -171,15 +174,15 @@ test('wrong kind', async t => {
   const { GreeterProxiable, GreeterV2Proxiable, ERC1967Proxy } = t.context;
 
   const impl = await GreeterProxiable.deploy();
-  await impl.deployed();
+  await impl.waitForDeployment();
   const proxy = await ERC1967Proxy.deploy(
-    impl.address,
+    await impl.getAddress(),
     getInitializerData(GreeterProxiable.interface, ['Hello, Hardhat!']),
   );
-  await proxy.deployed();
+  await proxy.waitForDeployment();
 
   // specify wrong kind
-  const greeter = await upgrades.forceImport(proxy.address, GreeterProxiable, { kind: 'transparent' });
+  const greeter = await upgrades.forceImport(await proxy.getAddress(), GreeterProxiable, { kind: 'transparent' });
   t.is(await greeter.greet(), 'Hello, Hardhat!');
 
   // an error is expected since the user force imported the wrong kind
@@ -191,14 +194,14 @@ test('import custom UUPS proxy', async t => {
   const { GreeterProxiable, GreeterV2Proxiable, CustomProxy } = t.context;
 
   const impl = await GreeterProxiable.deploy();
-  await impl.deployed();
+  await impl.waitForDeployment();
   const proxy = await CustomProxy.deploy(
-    impl.address,
+    await impl.getAddress(),
     getInitializerData(GreeterProxiable.interface, ['Hello, Hardhat!']),
   );
-  await proxy.deployed();
+  await proxy.waitForDeployment();
 
-  const greeter = await upgrades.forceImport(proxy.address, GreeterProxiable);
+  const greeter = await upgrades.forceImport(await proxy.getAddress(), GreeterProxiable);
   t.is(await greeter.greet(), 'Hello, Hardhat!');
 
   await upgrades.upgradeProxy(greeter, GreeterV2Proxiable);
@@ -208,14 +211,14 @@ test('import custom UUPS proxy with admin', async t => {
   const { GreeterProxiable, GreeterV2Proxiable, CustomProxyWithAdmin } = t.context;
 
   const impl = await GreeterProxiable.deploy();
-  await impl.deployed();
+  await impl.waitForDeployment();
   const proxy = await CustomProxyWithAdmin.deploy(
-    impl.address,
+    await impl.getAddress(),
     getInitializerData(GreeterProxiable.interface, ['Hello, Hardhat!']),
   );
-  await proxy.deployed();
+  await proxy.waitForDeployment();
 
-  const greeter = await upgrades.forceImport(proxy.address, GreeterProxiable);
+  const greeter = await upgrades.forceImport(await proxy.getAddress(), GreeterProxiable);
   t.is(await greeter.greet(), 'Hello, Hardhat!');
 
   await upgrades.upgradeProxy(greeter, GreeterV2Proxiable);
@@ -225,17 +228,17 @@ test('wrong implementation', async t => {
   const { Greeter, GreeterV2, ProxyAdmin, TransparentUpgradableProxy } = t.context;
 
   const impl = await Greeter.deploy();
-  await impl.deployed();
+  await impl.waitForDeployment();
   const admin = await ProxyAdmin.deploy();
-  await admin.deployed();
+  await admin.waitForDeployment();
   const proxy = await TransparentUpgradableProxy.deploy(
-    impl.address,
-    admin.address,
+    await impl.getAddress(),
+    await admin.getAddress(),
     getInitializerData(Greeter.interface, ['Hello, Hardhat!']),
   );
-  await proxy.deployed();
+  await proxy.waitForDeployment();
 
-  const greeter = await upgrades.forceImport(proxy.address, GreeterV2);
+  const greeter = await upgrades.forceImport(await proxy.getAddress(), GreeterV2);
   t.is(await greeter.greet(), 'Hello, Hardhat!');
 
   // since this is the wrong impl, expect it to have an error if using a non-existent function
@@ -247,26 +250,26 @@ test('multiple identical implementations', async t => {
   const { GreeterProxiable, GreeterV2Proxiable, ERC1967Proxy } = t.context;
 
   const impl = await GreeterProxiable.deploy();
-  await impl.deployed();
+  await impl.waitForDeployment();
   const proxy = await ERC1967Proxy.deploy(
-    impl.address,
+    await impl.getAddress(),
     getInitializerData(GreeterProxiable.interface, ['Hello, Hardhat!']),
   );
-  await proxy.deployed();
+  await proxy.waitForDeployment();
 
   const impl2 = await GreeterProxiable.deploy();
-  await impl2.deployed();
+  await impl2.waitForDeployment();
   const proxy2 = await ERC1967Proxy.deploy(
-    impl2.address,
+    await impl2.getAddress(),
     getInitializerData(GreeterProxiable.interface, ['Hello, Hardhat 2!']),
   );
-  await proxy2.deployed();
+  await proxy2.waitForDeployment();
 
-  const greeter = await upgrades.forceImport(proxy.address, GreeterProxiable);
+  const greeter = await upgrades.forceImport(await proxy.getAddress(), GreeterProxiable);
   const greeterUpgraded = await upgrades.upgradeProxy(greeter, GreeterV2Proxiable);
   t.is(await greeterUpgraded.greet(), 'Hello, Hardhat!');
 
-  const greeter2 = await upgrades.forceImport(proxy2.address, GreeterProxiable);
+  const greeter2 = await upgrades.forceImport(await proxy2.getAddress(), GreeterProxiable);
   const greeter2Upgraded = await upgrades.upgradeProxy(greeter2, GreeterV2Proxiable);
   t.is(await greeter2Upgraded.greet(), 'Hello, Hardhat 2!');
 });
@@ -275,23 +278,23 @@ test('same implementation', async t => {
   const { GreeterProxiable, ERC1967Proxy } = t.context;
 
   const impl = await GreeterProxiable.deploy();
-  await impl.deployed();
+  await impl.waitForDeployment();
   const proxy = await ERC1967Proxy.deploy(
-    impl.address,
+    await impl.getAddress(),
     getInitializerData(GreeterProxiable.interface, ['Hello, Hardhat!']),
   );
-  await proxy.deployed();
+  await proxy.waitForDeployment();
   const proxy2 = await ERC1967Proxy.deploy(
-    impl.address,
+    await impl.getAddress(),
     getInitializerData(GreeterProxiable.interface, ['Hello, Hardhat 2!']),
   );
-  await proxy2.deployed();
+  await proxy2.waitForDeployment();
 
-  const greeter = await upgrades.forceImport(proxy.address, GreeterProxiable);
-  const greeter2 = await upgrades.forceImport(proxy2.address, GreeterProxiable);
+  const greeter = await upgrades.forceImport(await proxy.getAddress(), GreeterProxiable);
+  const greeter2 = await upgrades.forceImport(await proxy2.getAddress(), GreeterProxiable);
 
-  const implAddr1 = await upgrades.erc1967.getImplementationAddress(greeter.address);
-  const implAddr2 = await upgrades.erc1967.getImplementationAddress(greeter2.address);
+  const implAddr1 = await upgrades.erc1967.getImplementationAddress(await greeter.getAddress());
+  const implAddr2 = await upgrades.erc1967.getImplementationAddress(await greeter2.getAddress());
   t.is(implAddr2, implAddr1);
 });
 
@@ -299,40 +302,41 @@ test('import transparents with different admin', async t => {
   const { Greeter, GreeterV2, ProxyAdmin, TransparentUpgradableProxy } = t.context;
 
   const impl = await Greeter.deploy();
-  await impl.deployed();
+  await impl.waitForDeployment();
   const admin = await ProxyAdmin.deploy();
-  await admin.deployed();
+  await admin.waitForDeployment();
   const proxy = await TransparentUpgradableProxy.deploy(
-    impl.address,
-    admin.address,
+    await impl.getAddress(),
+    await admin.getAddress(),
     getInitializerData(Greeter.interface, ['Hello, Hardhat!']),
   );
-  await proxy.deployed();
+  await proxy.waitForDeployment();
 
   const admin2 = await ProxyAdmin.deploy();
-  await admin2.deployed();
+  await admin2.waitForDeployment();
   const proxy2 = await TransparentUpgradableProxy.deploy(
-    impl.address,
-    admin2.address,
+    await impl.getAddress(),
+    await admin2.getAddress(),
     getInitializerData(Greeter.interface, ['Hello, Hardhat 2!']),
   );
-  await proxy2.deployed();
+  await proxy2.waitForDeployment();
 
-  const greeter = await upgrades.forceImport(proxy.address, Greeter);
-  const greeter2 = await upgrades.forceImport(proxy2.address, Greeter);
+  const greeter = await upgrades.forceImport(await proxy.getAddress(), Greeter);
+  const greeter2 = await upgrades.forceImport(await proxy2.getAddress(), Greeter);
 
   t.not(
-    await upgrades.erc1967.getAdminAddress(greeter2.address),
-    await upgrades.erc1967.getAdminAddress(greeter.address),
+    await upgrades.erc1967.getAdminAddress(await greeter2.getAddress()),
+    await upgrades.erc1967.getAdminAddress(await greeter.getAddress()),
   );
 
   // cannot upgrade directly
-  const e = await t.throwsAsync(() => upgrades.upgradeProxy(proxy.address, GreeterV2));
+  const proxyAddress = await proxy.getAddress();
+  const e = await t.throwsAsync(() => upgrades.upgradeProxy(proxyAddress, GreeterV2));
   t.is(NOT_REGISTERED_ADMIN, e.message, e.message);
 
   // prepare upgrades instead
-  const greeterV2ImplAddr = await upgrades.prepareUpgrade(greeter.address, GreeterV2);
-  const greeterV2ImplAddr_2 = await upgrades.prepareUpgrade(greeter2.address, GreeterV2);
+  const greeterV2ImplAddr = await upgrades.prepareUpgrade(await greeter.getAddress(), GreeterV2);
+  const greeterV2ImplAddr_2 = await upgrades.prepareUpgrade(await greeter2.getAddress(), GreeterV2);
 
   t.is(greeterV2ImplAddr_2, greeterV2ImplAddr);
 });