+++ packages/plugin-defender-hardhat/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Use `ethers` v6 and `hardhat-ethers` v3. ([#817](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/817))
+
 ## 1.9.0 (2023-06-14)
 
 - Update Defender Client dependencies. ([#806](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/806))