+++ packages/plugin-hardhat/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## Unreleased
+
+- Use `ethers` v6 and `hardhat-ethers` v3. ([#817](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/817))
+- Use `hardhat-verify` for proxy verification. ([#829](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/829))
+- Remove address override for deployments. ([#832](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/832))
+
+### Breaking changes
+This new major version requires `ethers` v6 and `@nomicfoundation/hardhat-ethers` v3 as peer dependencies.  
+For Etherscan verification, it also requires `@nomicfoundation/hardhat-verify`.
+
+### How to update from a previous version
+Update your existing project according to [Hardhat Toolbox v3's release notes](https://github.com/NomicFoundation/hardhat/releases/tag/%40nomicfoundation%2Fhardhat-toolbox%403.0.0).  
+Then update this plugin, for example: `npm install @openzeppelin/hardhat-upgrades@latest`
+
 ## 1.28.0 (2023-06-14)
 
 - **Breaking change**: Rename `walletId` option to `relayerId` for OpenZeppelin Platform deployments. Update dependencies. ([#806](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/806))